refactor(core): extract shared provider-pool and debug-dumper wiring helpers
Deduplicate ProviderConfigSnapshot construction across runner.rs/acp.rs/
daemon.rs and the DebugDumper::new wiring across those three plus
serve/agent_factory.rs into shared helpers in agent_setup.rs, completing
the provider-pool-wiring slice left open by the apply_common_tool_gating
extraction. Pure refactor, no behavior change.

Closes #5610
