+++ src/daemon.rs
@@ -723,46 +723,7 @@ pub(crate) async fn run_daemon(
     // #5450: built here, where the full `Config` is still in scope — mirrors
     // `src/runner.rs`'s CLI-path snapshot construction, so the daemon's agent gets a populated
     // `provider_pool` too (previously left empty, breaking `resolve_background_provider`).
-    let provider_config_snapshot = zeph_core::ProviderConfigSnapshot {
-        claude_api_key: config
-            .secrets
-            .claude_api_key
-            .as_ref()
-            .map(|s| s.expose().to_owned()),
-        openai_api_key: config
-            .secrets
-            .openai_api_key
-            .as_ref()
-            .map(|s| s.expose().to_owned()),
-        gemini_api_key: config
-            .secrets
-            .gemini_api_key
-            .as_ref()
-            .map(|s| s.expose().to_owned()),
-        compatible_api_keys: config
-            .secrets
-            .compatible_api_keys
-            .iter()
-            .map(|(k, v)| (k.clone(), v.expose().to_owned()))
-            .collect(),
-        llm_request_timeout_secs: config.timeouts.llm_request_timeout_secs,
-        embedding_model: config.llm.embedding_model.clone(),
-        gonka_private_key: config
-            .secrets
-            .gonka_private_key
-            .as_ref()
-            .map(|s| zeroize::Zeroizing::new(s.expose().to_owned())),
-        gonka_address: config
-            .secrets
-            .gonka_address
-            .as_ref()
-            .map(|s| s.expose().to_owned()),
-        cocoon_access_hash: config
-            .secrets
-            .cocoon_access_hash
-            .as_ref()
-            .map(|s| s.expose().to_owned()),
-    };
+    let provider_config_snapshot = agent_setup::build_provider_config_snapshot(config);
 
     let (loopback_channel, loopback_handle) = zeph_core::LoopbackChannel::pair(64);
 
@@ -976,13 +937,12 @@ pub(crate) async fn run_daemon(
     // `spawn_gateway_server` only forwards webhooks into an already-built agent) never wired
     // `[debug]`, unlike the CLI (src/runner.rs) and ACP (src/acp.rs) agent-construction paths.
     if config.debug.enabled {
-        match zeph_core::debug_dump::DebugDumper::new(
+        agent = agent_setup::apply_debug_dumper(
+            agent,
             config.debug.output_dir.as_path(),
             config.debug.format,
-        ) {
-            Ok(dumper) => agent = agent.with_debug_dumper(dumper),
-            Err(e) => tracing::warn!(error = %e, "debug dump initialization failed"),
-        }
+        )
+        .0;
     }
 
     agent.load_history().await?;
